+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.mockito"
+        applicationId "com.example.tests"
         minSdk 23
         targetSdk 32
         versionCode 1
