--- app/src/main/java/com/example/mockito/repository/GitHubService.kt
@@ -1,18 +1,0 @@
-package com.example.mockito.repository
-
-import com.example.mockito.tests_search.model.SearchResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Headers
-import retrofit2.http.Query
-
-/**
- * Документация https://developer.github.com/v3/search/
- */
-
-interface GitHubService {
-
-    @Headers("Accept: application/vnd.github.mercy-preview+json")
-    @GET("search/repositories")
-    fun searchGithub(@Query("q") term: String?): Call<SearchResponse?>?
-}
