--- app/src/main/java/com/example/mockito/repository/RepositoryContract.kt
@@ -1,6 +1,0 @@
-package com.example.mockito.repository
-
-interface RepositoryContract {
-
-    fun searchGithub(query: String, callback: RepositoryCallback)
-}
