--- app/src/main/java/com/example/mockito/tests_details/DetailsActivity.kt
@@ -1,55 +1,0 @@
-package com.example.mockito.tests_details
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.mockito.R
-import com.example.mockito.databinding.ActivityDetailsBinding
-import java.util.Locale
-
-/**
- * Экран (DetailsActivity), на котором мы и сделаем первые шаги в использовании Роболектрика.
- * На этом экране мы будем отображать количество найденных аккаунтов по запросу.
- */
-class DetailsActivity : AppCompatActivity(), ViewDetailsContract {
-
-    private lateinit var binding: ActivityDetailsBinding
-
-    private val presenter: PresenterDetailsContract = DetailsPresenter(this)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityDetailsBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        setUI()
-    }
-
-    private fun setUI() {
-        val count = intent.getIntExtra(TOTAL_COUNT_EXTRA, 0)
-        presenter.setCounter(count)
-        setCountText(count)
-        binding.decrementButton.setOnClickListener { presenter.onDecrement() }
-        binding.incrementButton.setOnClickListener { presenter.onIncrement() }
-    }
-
-    override fun setCount(count: Int) {
-        setCountText(count)
-    }
-
-    private fun setCountText(count: Int) {
-        binding.totalCountTextView.text =
-            String.format(Locale.getDefault(), getString(R.string.results_count), count)
-    }
-
-    companion object {
-
-        const val TOTAL_COUNT_EXTRA = "TOTAL_COUNT_EXTRA"
-
-        fun getIntent(context: Context, totalCount: Int): Intent {
-            return Intent(context, DetailsActivity::class.java).apply {
-                putExtra(TOTAL_COUNT_EXTRA, totalCount)
-            }
-        }
-    }
-}
