--- app/src/main/java/com/example/mockito/tests_details/PresenterDetailsContract.kt
@@ -1,10 +1,0 @@
-package com.example.mockito.tests_details
-
-import com.example.mockito.presenter.PresenterContract
-
-internal interface PresenterDetailsContract : PresenterContract {
-
-    fun setCounter(count: Int)
-    fun onIncrement()
-    fun onDecrement()
-}
