--- app/src/main/java/com/example/mockito/tests_details/ViewDetailsContract.kt
@@ -1,8 +1,0 @@
-package com.example.mockito.tests_details
-
-import com.example.mockito.view.ViewContract
-
-internal interface ViewDetailsContract : ViewContract {
-
-    fun setCount(count: Int)
-}
