--- app/src/main/java/com/example/mockito/tests_search/PresenterSearchContract.kt
@@ -1,8 +1,0 @@
-package com.example.mockito.tests_search
-
-import com.example.mockito.presenter.PresenterContract
-
-internal interface PresenterSearchContract : PresenterContract {
-
-    fun searchGitHub(searchQuery: String)
-}
