--- app/src/main/java/com/example/mockito/tests_search/ViewSearchContract.kt
@@ -1,16 +1,0 @@
-package com.example.mockito.tests_search
-
-import com.example.mockito.tests_search.model.SearchResult
-import com.example.mockito.view.ViewContract
-
-internal interface ViewSearchContract : ViewContract {
-
-    fun displaySearchResults(
-        searchResults: List<SearchResult>,
-        totalCount: Int
-    )
-
-    fun displayError()
-    fun displayError(error: String)
-    fun displayLoading(show: Boolean)
-}
