--- app/src/main/java/com/example/mockito/tests_search/model/SearchResponse.kt
@@ -1,13 +1,0 @@
-package com.example.mockito.tests_search.model
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-data class SearchResponse(
-    @SerializedName("total_count")
-    @Expose
-    val totalCount: Int?,
-    @SerializedName("items")
-    @Expose
-    val searchResults: List<SearchResult>?
-)
