--- app/src/main/java/com/example/mockito/view/ViewContract.kt
@@ -1,6 +1,0 @@
-package com.example.mockito.view
-
-/**
- * общий для всех экранов интерфейс контракта view
- */
-internal interface ViewContract {}
