+++ app/src/main/java/com/example/tests/presenter/PresenterContract.kt
@@ -1,0 +1,14 @@
+package com.example.tests.presenter
+
+import com.example.tests.view.ViewContract
+
+/**
+ * Общий для всех экранов интерфейс с функциями onAttach() и onDetach()
+ * Помним, что когда тело функции пустое, то можем эту функцию далее не переопределять
+ */
+internal interface PresenterContract {
+
+    fun onAttach(view: ViewContract) {}
+
+    fun onDetach() {}
+}
