+++ app/src/main/java/com/example/tests/repository/GitHubRepository.kt
@@ -1,0 +1,32 @@
+package com.example.tests.repository
+
+import com.example.tests.tests_search.model.SearchResponse
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+
+class GitHubRepository(private val gitHubService: GitHubService) : RepositoryContract {
+
+    override fun searchGithub(
+        query: String,
+        callback: RepositoryCallback
+    ) {
+        val call = gitHubService.searchGithub(query)
+        call?.enqueue(object : Callback<SearchResponse?> {
+
+            override fun onResponse(
+                call: Call<SearchResponse?>,
+                response: Response<SearchResponse?>
+            ) {
+                callback.handleGitHubResponse(response)
+            }
+
+            override fun onFailure(
+                call: Call<SearchResponse?>,
+                t: Throwable
+            ) {
+                callback.handleGitHubError()
+            }
+        })
+    }
+}
