+++ app/src/main/java/com/example/tests/repository/RepositoryCallback.kt
@@ -1,0 +1,10 @@
+package com.example.tests.repository
+
+import com.example.tests.tests_search.model.SearchResponse
+import retrofit2.Response
+
+interface RepositoryCallback {
+
+    fun handleGitHubResponse(response: Response<SearchResponse?>?)
+    fun handleGitHubError()
+}
