+++ app/src/main/java/com/example/tests/repository/RepositoryContract.kt
@@ -1,0 +1,6 @@
+package com.example.tests.repository
+
+interface RepositoryContract {
+
+    fun searchGithub(query: String, callback: RepositoryCallback)
+}
