+++ app/src/main/java/com/example/tests/tests_details/PresenterDetailsContract.kt
@@ -1,0 +1,10 @@
+package com.example.tests.tests_details
+
+import com.example.tests.presenter.PresenterContract
+
+internal interface PresenterDetailsContract : PresenterContract {
+
+    fun setCounter(count: Int)
+    fun onIncrement()
+    fun onDecrement()
+}
