+++ app/src/main/java/com/example/tests/tests_details/ViewDetailsContract.kt
@@ -1,0 +1,8 @@
+package com.example.tests.tests_details
+
+import com.example.tests.view.ViewContract
+
+internal interface ViewDetailsContract : ViewContract {
+
+    fun setCount(count: Int)
+}
