+++ app/src/main/java/com/example/tests/tests_search/PresenterSearchContract.kt
@@ -1,0 +1,8 @@
+package com.example.tests.tests_search
+
+import com.example.tests.presenter.PresenterContract
+
+internal interface PresenterSearchContract : PresenterContract {
+
+    fun searchGitHub(searchQuery: String)
+}
