+++ app/src/main/java/com/example/tests/tests_search/SearchPresenter.kt
@@ -1,0 +1,64 @@
+package com.example.tests.tests_search
+
+import android.util.Log
+import com.example.tests.repository.RepositoryCallback
+import com.example.tests.repository.RepositoryContract
+import com.example.tests.tests_search.model.SearchResponse
+import retrofit2.Response
+
+/**
+ * В архитектуре MVP все запросы на получение данных адресуются в Репозиторий.
+ * Запросы могут проходить через Interactor или UseCase, использовать источники
+ * данных (DataSource), но суть от этого не меняется.
+ * Непосредственно Презентер отвечает за управление потоками запросов и ответов,
+ * выступая в роли регулировщика движения на перекрестке.
+ */
+
+internal class SearchPresenter internal constructor(
+    private val viewContract: ViewSearchContract,
+    private val repository: RepositoryContract
+) : PresenterSearchContract, RepositoryCallback {
+
+    /*override fun onAttach(context: Context) {
+        super.onAttach(context)
+
+        if (context !is PresenterContract) {
+            throw IllegalStateException("Activity must implement PresenterContract")
+        }
+    }*/
+
+    override fun searchGitHub(searchQuery: String) {
+        viewContract.displayLoading(true)
+        repository.searchGithub(searchQuery, this)
+    }
+
+    override fun handleGitHubResponse(response: Response<SearchResponse?>?) {
+        viewContract.displayLoading(false)
+        if (response != null && response.isSuccessful) {
+            val searchResponse = response.body()
+            val searchResults = searchResponse?.searchResults
+            val totalCount = searchResponse?.totalCount
+            if (searchResults != null && totalCount != null) {
+                viewContract.displaySearchResults(
+                    searchResults,
+                    totalCount
+                )
+            } else {
+                viewContract.displayError("Search results or total count are null")
+            }
+        } else {
+            viewContract.displayError("Response is null or unsuccessful")
+        }
+    }
+
+    override fun handleGitHubError() {
+        viewContract.displayLoading(false)
+        viewContract.displayError()
+    }
+
+    override fun onDetach() {
+        Log.d("MainActivity", "onDetach")
+
+        super.onDetach()
+    }
+}
