+++ app/src/main/java/com/example/tests/tests_search/ViewSearchContract.kt
@@ -1,0 +1,16 @@
+package com.example.tests.tests_search
+
+import com.example.tests.tests_search.model.SearchResult
+import com.example.tests.view.ViewContract
+
+internal interface ViewSearchContract : ViewContract {
+
+    fun displaySearchResults(
+        searchResults: List<SearchResult>,
+        totalCount: Int
+    )
+
+    fun displayError()
+    fun displayError(error: String)
+    fun displayLoading(show: Boolean)
+}
