+++ app/src/main/java/com/example/tests/tests_search/model/SearchResponse.kt
@@ -1,0 +1,13 @@
+package com.example.tests.tests_search.model
+
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+
+data class SearchResponse(
+    @SerializedName("total_count")
+    @Expose
+    val totalCount: Int?,
+    @SerializedName("items")
+    @Expose
+    val searchResults: List<SearchResult>?
+)
