+++ app/src/main/java/com/example/tests/view/ViewContract.kt
@@ -1,0 +1,6 @@
+package com.example.tests.view
+
+/**
+ * общий для всех экранов интерфейс контракта view
+ */
+internal interface ViewContract {}
